All requested changes were implemented.
