+++ WordOfTheDay.PerformanceCheck/WordOfTheDay.PerformanceCheck/Word.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace WordOfTheDay.PerformanceCheck
+{
+    public class Word
+    {
+        public string Email { get; set; }
+        public string Text { get; set; }
+    }
+}
