+++ WordOfTheDay.PerformanceCheck/WordOfTheDay.PerformanceCheck/WordOfTheDay.PerformanceCheck.csproj
@@ -5,8 +5,4 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\WordOfTheDay.Repository\WordOfTheDay.Repository.csproj" />
-  </ItemGroup>
-
 </Project>
